Replace arr.ptp() with np.ptp(arr) and fix float type issues for JSON serialization (#61)
* Replaced deprecated np.ndarray.ptp method with np.ptp function

* Ensure dz_total is float64 for consistent JSON serialization

+++ demcoreg/dem_align.py
@@ -409,7 +409,7 @@ def main(argv=None):
                 #Want to have max_tilt check here
                 #max_tilt = 4.0 #m
                 #Should do percentage
-                #vals.ptp() > max_tilt
+                #np.ptp(vals) > max_tilt
 
                 #Note: dimensions of ds and vals will be different as vals are computed for clipped intersection
                 #Need to recompute planar offset for full src_dem_ds_align extent and apply
@@ -545,7 +545,7 @@ def main(argv=None):
         align_stats['res']['coreg'] = res
         align_stats['center_coord'] = {'lon':center_coord_ll[0], 'lat':center_coord_ll[1], \
                 'x':center_coord_xy[0], 'y':center_coord_xy[1]}
-        align_stats['shift'] = {'dx':dx_total, 'dy':dy_total, 'dz':dz_total, 'dm':dm_total}
+        align_stats['shift'] = {'dx':dx_total, 'dy':dy_total, 'dz':np.float64(dz_total), 'dm':dm_total}
         #This tiltcorr flag gets set to false, need better flag
         if tiltcorr:
             align_stats['tiltcorr'] = {}
